Add unique id to each item in the basket
Assuming that items can be added more than once to a basket, this means
we need a way of referring to a specific item so we can remove that one
item.

Using unique ids instead of storing a count because that doesn't play
nice with distributed systems, multiple things could try to change the
count and overwrite each other. Where as this way deals with immutable
objects which play much nicer together.