Add authentication with JWT
In order to store baskets against users, we can get the current user
from the session. JWT seems like the most sensible approach as it leaves
us with a stateless server. Note that tests will have to mock their own
middleware to add the user object.

I am using keycloak to test this, need to automate the setup of that in
the future.